BUG: Search results with >10k results can not all be discovered
The maximum offset value possible is 9500, so only the first 10k
results can be listed for searches with more results than this.

We now automatically detect when results are repeated for this
reason by comparing the first record in the listing for consecutive
offset values. This solution is a graceful degredation - if pangaea
changes their API to support larger offset values in the future,
more results will be discovered and returned. A warning is issued
instead of an error being raised.