JNB: Investigate extensions in dataset
This code block was actually run before restricting to just image
extensions, but after removing blacklisted subdomains. However
it doesn't make sense to do the subdomain blacklist investigation
without restricting the URLs to just be images. So we pick an
order which makes sense for subdomain investigation, but not
really for file extensions.