ENH: Rate-limit download requests to once every 0.167 seconds
The request rate which pangaea will serve is rate limited to a maximum
of 180 per 30s for each client connecting to it. If we exceed this, we
get a HTTP 403 error. The error message passed on from pangaeapy reads:

You started too many requests per time interval to this web application.
A client is only allowed to start a maximum of 180 requests per 30s to
this service. Please try again later!

In this commit, we throttle ourselves to prevent this error.